Adds media style for select box width (#148)
For mobile screen sizes on the submission add cross step, this stops
the select box from bleeding off the right of the screen.

This is more of a stop gap measure. A select is not really appropriate
for the cross list on mobile. A better solution to this should be
found.

ARXIVNG-2200