Fixes bug with can_procede_to()
This had an OR clause that started with self.is_done(previous_stage)
but that is not correct.
ex.
1 Start a submission,
2 Go throught each stage until finalize, all stages are now seen
3 Go back to FileUpload and delete all files
4 Skip to finalize stage: should not be able to

but.. this clause will return true since the previous stage is Opt_Metadata(required=False, must_see=True)
and is_done returns True.

The second clause in the OR is also incorrect in the case that
previous_stage is not required but is must_see.