Refactored flow_control to not use exceptions
flow_control was using BadRequest exceptions to make non-local jumps
from the controllers to locations in the flow_control function.

Using exceptions in this way and incorrectly using BadRequest lead to
confusing code.

This is was probably due to the confusing way flow_control and
routes.ui.handle are used with a Flask Blueprint.