Fixes recent problem with process step.
This fixes the config of the preview service. The config variables
for the preview service changed from PREVIEW_ENDPOINT to
SUBMISSION_PREVIEW_ENDPOINT. This caused a problem because the name
of config variables is fixed by the name of the integration class.

See arxiv-base/base/arxiv/integration/api/service.py L250 for where
the service configuration keys are created using the class name.

The other problem was the refresh was no longer getting into the head
of the HTML page for the process step.